+++ demos/employees/src/main/java/is/codion/framework/demos/employees/testing/EmployeesServletLoadTest.java
@@ -135,7 +135,7 @@ public final class EmployeesServletLoadTest {
 			client.connection().insert(client.entities().builder(Employee.TYPE)
 							.with(Employee.DEPARTMENT_FK, department)
 							.with(Employee.NAME, randomString(8))
-							.with(Employee.JOB, Employee.JOB_VALUES.get(random.nextInt(Employee.JOB_VALUES.size())).get())
+							.with(Employee.JOB, Employee.JOB_ITEMS.get(random.nextInt(Employee.JOB_ITEMS.size())).value())
 							.with(Employee.SALARY, BigDecimal.valueOf(random.nextInt(1000) + 1000))
 							.with(Employee.HIREDATE, LocalDate.now())
 							.with(Employee.COMMISSION, random.nextDouble() * 500)