+++ swing/common-model/src/main/java/is/codion/swing/common/model/component/combobox/ItemComboBoxModel.java
@@ -62,7 +62,7 @@ public final class ItemComboBoxModel<T> extends DefaultFilterComboBoxModel<Item<
 	 */
 	public int indexOf(T value) {
 		for (int i = 0; i < getSize(); i++) {
-			if (Objects.equals(getElementAt(i).get(), value)) {
+			if (Objects.equals(getElementAt(i).value(), value)) {
 				return i;
 			}
 		}