+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/dialog/DefaultFontSizeSelectionDialogBuilder.java
@@ -104,7 +104,7 @@ final class DefaultFontSizeSelectionDialogBuilder implements FontSizeSelectionDi
 		}
 
 		private int selectedFontSize() {
-			return fontSizeComboBoxModel.selectedValue().get();
+			return fontSizeComboBoxModel.selectedValue().value();
 		}
 
 		private static List<Item<Integer>> initializeValues() {
@@ -133,7 +133,7 @@ final class DefaultFontSizeSelectionDialogBuilder implements FontSizeSelectionDi
 				Component component = defaultListCellRenderer.getListCellRendererComponent(list, value, index, isSelected, cellHasFocus);
 				if (index >= 0) {
 					Font font = component.getFont();
-					int newSize = Math.round(font.getSize() * (values.get(index).get() / (float) currentFontSize.doubleValue()));
+					int newSize = Math.round(font.getSize() * (values.get(index).value() / (float) currentFontSize.doubleValue()));
 					component.setFont(new Font(font.getName(), font.getStyle(), newSize));
 				}