+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntitySearchField.java
@@ -652,7 +652,7 @@ public final class EntitySearchField extends HintTextField {
 			}
 			JPanel columnBasePanel = columnBasePanelBuilder.build();
 			if (columnComboBoxModel.getSize() > 0) {
-				columnComboBoxModel.selectionChanged().addConsumer(selected -> cardLayout.show(columnBasePanel, selected.get().name()));
+				columnComboBoxModel.selectionChanged().addConsumer(selected -> cardLayout.show(columnBasePanel, selected.value().name()));
 				columnComboBoxModel.setSelectedItem(columnComboBoxModel.getElementAt(0));
 			}